requirements.txt
requests
clubhouse-py
bottle